facade part2
